Refactor Standley Lake Screen into WindStationTab component
- Moved all logic and UI from StandleyLakeScreen to a new WindStationTab component.
- Introduced WindStationData and WindStationConfig types for better data handling.
- Created utility functions for wind data processing in windStationUtils.
- Updated StandleyLakeScreen to utilize the new WindStationTab component.
- Enhanced data freshness messaging and transmission quality indicators.
- Improved code organization and readability.
