+++ app/(tabs)/standley-lake.tsx
@@ -1,474 +1,17 @@
-import { ThemedText } from '@/components/ThemedText';
-import { ThemedView } from '@/components/ThemedView';
-import { CustomWindChart } from '@/components/CustomWindChart';
-import { useStandleyLakeWind } from '@/hooks/useStandleyLakeWind';
-import { useThemeColor } from '@/hooks/useThemeColor';
-import { getWindChartTimeWindow } from '@/utils/timeWindowUtils';
 import React from 'react';
-import {
-  ActivityIndicator,
-  RefreshControl,
-  ScrollView,
-  StyleSheet,
-  TouchableOpacity,
-  View
-} from 'react-native';
-import { useFocusEffect } from '@react-navigation/native';
+import WindStationTab from '@/components/WindStationTab';
+import { useStandleyLakeWind } from '@/hooks/useStandleyLakeWind';
 
 export default function StandleyLakeScreen() {
-  const {
-    windData,
-    chartData,
-    currentConditions,
-    analysis,
-    isLoading,
-    error,
-    currentConditionsUpdated,
-    refreshData,
-    refreshCurrentConditions,
-    clearCache
-  } = useStandleyLakeWind();
-
-  // const textColor = useThemeColor({}, 'text'); // Currently unused
-  // const backgroundColor = useThemeColor({}, 'background'); // Currently unused
-  const tintColor = useThemeColor({}, 'tint');
-  const cardColor = useThemeColor({}, 'card');
-
-  // Track if we've loaded data initially
-  const [hasInitiallyLoaded, setHasInitiallyLoaded] = React.useState(false);
-
-  // Only refresh data on first navigation to the tab
-  useFocusEffect(
-    React.useCallback(() => {
-      if (!hasInitiallyLoaded) {
-        console.log('🏔️ Standley Lake tab focused for first time - loading data...');
-        refreshData();
-        refreshCurrentConditions();
-        setHasInitiallyLoaded(true);
-      }
-    }, [hasInitiallyLoaded, refreshData, refreshCurrentConditions])
-  );
-
-  const handleRefresh = async () => {
-    await refreshData();
-  };
-
-  // Helper functions for current conditions
-  const getCurrentWindSpeed = () => {
-    return currentConditions?.windSpeedMph || null;
-  };
-
-  const getCurrentWindDirection = () => {
-    return currentConditions?.windDirection;
-  };
-
-  const getWindDirectionText = (degrees: number | undefined): string => {
-    if (degrees === undefined || degrees === null) return '--';
-    
-    const directions = ['N', 'NNE', 'NE', 'ENE', 'E', 'ESE', 'SE', 'SSE', 'S', 'SSW', 'SW', 'WSW', 'W', 'WNW', 'NW', 'NNW'];
-    const index = Math.round(degrees / 22.5) % 16;
-    return directions[index];
-  };
-
-  const formatCurrentConditionsUpdated = () => {
-    // Prioritize real-time data timestamp if available
-    if (currentConditions && currentConditionsUpdated) {
-      return currentConditionsUpdated.toLocaleTimeString('en-US', { 
-        hour: 'numeric', 
-        minute: '2-digit', 
-        hour12: true 
-      });
+  const data = useStandleyLakeWind();
+  
+  const config = {
+    name: 'Standley Lake Wind Monitor',
+    subtitle: 'Real-time wind conditions',
+    features: {
+      // Standley Lake doesn't have transmission quality monitoring or external link
     }
-    
-    // Fall back to latest historical data timestamp
-    if (windData.length > 0) {
-      const latest = windData[windData.length - 1];
-      const latestTime = new Date(latest.time);
-      return latestTime.toLocaleTimeString('en-US', { 
-        hour: 'numeric', 
-        minute: '2-digit', 
-        hour12: true 
-      });
-    }
-    
-    return 'Never';
   };
 
-  // Check if data is stale - prioritize real-time data for current conditions
-  const isDataStale = () => {
-    // Check real-time data first - if we have recent real-time data, consider it fresh
-    if (currentConditions && currentConditionsUpdated) {
-      const now = new Date();
-      const thirtyMinutesAgo = new Date(now.getTime() - 30 * 60 * 1000);
-      
-      if (currentConditionsUpdated > thirtyMinutesAgo) {
-        return false; // Real-time data is fresh
-      }
-    }
-    
-    // If no real-time data, check historical data  
-    if (windData.length === 0) return true;
-    
-    const latest = windData[windData.length - 1];
-    const latestTime = new Date(latest.time);
-    const now = new Date();
-    const thirtyMinutesAgo = new Date(now.getTime() - 30 * 60 * 1000);
-    
-    // Consider historical data fresh if it's within 30 minutes
-    return latestTime < thirtyMinutesAgo;
-  };
-
-  const getDataFreshnessMessage = () => {
-    // Check real-time data first
-    if (currentConditions && currentConditionsUpdated) {
-      const now = new Date();
-      const diffMinutes = Math.floor((now.getTime() - currentConditionsUpdated.getTime()) / 60000);
-      
-      if (diffMinutes === 0) return 'Real-time data current';
-      if (diffMinutes < 5) return `Real-time data from ${diffMinutes} minute${diffMinutes === 1 ? '' : 's'} ago`;
-      if (diffMinutes < 30) return `Real-time data from ${diffMinutes} minutes ago`;
-      
-      // If real-time data is older than 30 minutes, fall through to check historical data
-    }
-    
-    // Fall back to historical data messaging
-    if (windData.length === 0) return 'No data available';
-    
-    const latest = windData[windData.length - 1];
-    const latestTime = new Date(latest.time);
-    const now = new Date();
-    const diffMinutes = Math.floor((now.getTime() - latestTime.getTime()) / 60000);
-    
-    if (diffMinutes < 30) return 'Historical data is current';
-    if (diffMinutes < 120) return `Last historical reading ${diffMinutes} minutes ago`;
-    
-    const diffHours = Math.floor(diffMinutes / 60);
-    return `⚠️ Station may be offline - last reading ${diffHours} hours ago`;
-  };
-
-  return (
-    <ThemedView style={styles.container}>
-      <ScrollView
-        contentInsetAdjustmentBehavior="automatic"
-        refreshControl={
-          <RefreshControl
-            refreshing={isLoading}
-            onRefresh={handleRefresh}
-            tintColor={tintColor}
-          />
-        }
-      >
-        <View style={styles.header}>
-          <View>
-            <ThemedText type="title">Standley Lake Wind Monitor</ThemedText>
-            <ThemedText style={styles.subtitle}>Real-time wind conditions</ThemedText>
-          </View>
-        </View>
-
-        {error && (
-          <View style={styles.errorContainer}>
-            <ThemedText style={styles.errorText}>⚠️ {error}</ThemedText>
-            <TouchableOpacity style={styles.retryButton} onPress={handleRefresh}>
-              <ThemedText style={[styles.retryButtonText, { color: tintColor }]}>
-                Retry
-              </ThemedText>
-            </TouchableOpacity>
-          </View>
-        )}
-
-        {isLoading && (
-          <View style={styles.loadingContainer}>
-            <ActivityIndicator size="large" color={tintColor} />
-            <ThemedText style={styles.loadingText}>Updating wind data...</ThemedText>
-          </View>
-        )}
-
-        {/* Data Freshness Alert */}
-        {windData.length > 0 && isDataStale() && (
-          <View style={styles.staleDataContainer}>
-            <ThemedText style={styles.staleDataText}>
-              {getDataFreshnessMessage()}
-            </ThemedText>
-            <ThemedText style={[styles.staleDataSubtext]}>
-              Station may not be reporting. Try refreshing or check back later.
-            </ThemedText>
-          </View>
-        )}
-
-        {/* Current Conditions */}
-        <View style={[styles.currentConditionsCard, { backgroundColor: cardColor }]}>
-          <ThemedText type="subtitle" style={styles.cardTitle}>Current Conditions</ThemedText>
-          <View style={styles.currentConditionsGrid}>
-            <View style={styles.conditionItem}>
-              <ThemedText style={styles.conditionLabel}>Wind Speed</ThemedText>
-              <ThemedText style={styles.conditionValue}>
-                {getCurrentWindSpeed()?.toFixed(1) || '--'} mph
-              </ThemedText>
-            </View>
-            <View style={styles.conditionItem}>
-              <ThemedText style={styles.conditionLabel}>Direction</ThemedText>
-              <ThemedText style={styles.conditionValue}>
-                {getWindDirectionText(getCurrentWindDirection())} ({getCurrentWindDirection()?.toFixed(0) || '--'}°)
-              </ThemedText>
-            </View>
-            <View style={styles.conditionItem}>
-              <ThemedText style={styles.conditionLabel}>Last Updated</ThemedText>
-              <ThemedText style={styles.conditionValue}>
-                {formatCurrentConditionsUpdated()}
-              </ThemedText>
-            </View>
-          </View>
-          {/* Data freshness message */}
-          <ThemedText style={styles.freshnessMessage}>
-            {getDataFreshnessMessage()}
-          </ThemedText>
-        </View>
-
-        {/* Wind Chart */}
-        {chartData.length > 0 ? (
-          <View style={[styles.chartCard, { backgroundColor: cardColor }]}>
-            <CustomWindChart
-              data={chartData}
-              title="Today's Wind Speed"
-              timeWindow={getWindChartTimeWindow()}
-            />
-          </View>
-        ) : (
-          <View style={[styles.noDataCard, { backgroundColor: cardColor }]}>
-            <ThemedText style={styles.noDataText}>
-              No wind data available for today. Pull to refresh.
-            </ThemedText>
-          </View>
-        )}
-
-        {/* Historic Data Points */}
-        {windData.length > 0 && (
-          <View style={[styles.dataPointsCard, { backgroundColor: cardColor }]}>
-            <ThemedText style={styles.dataPointsText}>
-              Historic Data Points: {windData.length}
-            </ThemedText>
-          </View>
-        )}
-
-        {/* Recent Wind Analysis */}
-        {analysis && (
-          <View style={[styles.analysisCard, { backgroundColor: cardColor }]}>
-            <ThemedText type="subtitle" style={styles.cardTitle}>Recent Wind Analysis (Last Hour)</ThemedText>
-            <View style={styles.analysisGrid}>
-              <View style={styles.analysisItem}>
-                <ThemedText style={styles.analysisValue}>
-                  {analysis.averageSpeed.toFixed(1)} mph
-                </ThemedText>
-                <ThemedText style={styles.analysisLabel}>Avg Speed</ThemedText>
-              </View>
-              <View style={styles.analysisItem}>
-                <ThemedText style={styles.analysisValue}>
-                  {analysis.directionConsistency.toFixed(0)}%
-                </ThemedText>
-                <ThemedText style={styles.analysisLabel}>Direction Consistency</ThemedText>
-              </View>
-              <View style={styles.analysisItem}>
-                <ThemedText style={styles.analysisValue}>
-                  {analysis.consecutiveGoodPoints}
-                </ThemedText>
-                <ThemedText style={styles.analysisLabel}>Good Points</ThemedText>
-              </View>
-            </View>
-            <ThemedText style={styles.analysisDescription}>
-              {analysis.analysis}
-            </ThemedText>
-          </View>
-        )}
-
-        {/* Debug Actions */}
-        {__DEV__ && (
-          <View style={styles.debugContainer}>
-            <TouchableOpacity
-              style={[styles.debugButton, { borderColor: tintColor }]}
-              onPress={clearCache}
-            >
-              <ThemedText style={[styles.debugButtonText, { color: tintColor }]}>
-                Clear Cache
-              </ThemedText>
-            </TouchableOpacity>
-            <TouchableOpacity
-              style={[styles.debugButton, { borderColor: tintColor, marginTop: 8 }]}
-              onPress={refreshCurrentConditions}
-            >
-              <ThemedText style={[styles.debugButtonText, { color: tintColor }]}>
-                Test Real-Time API
-              </ThemedText>
-            </TouchableOpacity>
-          </View>
-        )}
-      </ScrollView>
-    </ThemedView>
-  );
+  return <WindStationTab data={data} config={config} />;
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-  },
-  header: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    paddingHorizontal: 16,
-    paddingTop: 16,
-    paddingBottom: 8,
-  },
-  subtitle: {
-    fontSize: 16,
-    opacity: 0.7,
-    marginTop: 4,
-  },
-  errorContainer: {
-    margin: 16,
-    padding: 16,
-    backgroundColor: 'rgba(255, 59, 48, 0.1)',
-    borderRadius: 12,
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-  },
-  errorText: {
-    color: '#FF3B30',
-    flex: 1,
-  },
-  retryButton: {
-    marginLeft: 12,
-  },
-  retryButtonText: {
-    fontWeight: '600',
-  },
-  loadingContainer: {
-    alignItems: 'center',
-    padding: 32,
-  },
-  loadingText: {
-    marginTop: 12,
-    opacity: 0.7,
-  },
-  currentConditionsCard: {
-    margin: 16,
-    padding: 16,
-    borderRadius: 12,
-  },
-  cardTitle: {
-    marginBottom: 12,
-  },
-  currentConditionsGrid: {
-    flexDirection: 'row',
-    flexWrap: 'wrap',
-    justifyContent: 'space-between',
-  },
-  conditionItem: {
-    width: '48%',
-    marginBottom: 12,
-  },
-  conditionLabel: {
-    fontSize: 12,
-    opacity: 0.7,
-    marginBottom: 4,
-  },
-  conditionValue: {
-    fontSize: 18,
-    fontWeight: '600',
-  },
-  chartCard: {
-    margin: 16,
-    borderRadius: 12,
-    overflow: 'hidden',
-  },
-  dataPointsCard: {
-    margin: 16,
-    marginTop: 8,
-    padding: 12,
-    borderRadius: 8,
-    alignItems: 'center',
-  },
-  dataPointsText: {
-    fontSize: 14,
-    opacity: 0.8,
-  },
-  noDataCard: {
-    margin: 16,
-    padding: 32,
-    borderRadius: 12,
-    alignItems: 'center',
-  },
-  noDataText: {
-    textAlign: 'center',
-    opacity: 0.7,
-  },
-  analysisCard: {
-    margin: 16,
-    padding: 16,
-    borderRadius: 12,
-  },
-  analysisGrid: {
-    flexDirection: 'row',
-    justifyContent: 'space-around',
-    marginBottom: 16,
-  },
-  analysisItem: {
-    alignItems: 'center',
-    flex: 1,
-  },
-  analysisValue: {
-    fontSize: 20,
-    fontWeight: '700',
-    marginBottom: 4,
-  },
-  analysisLabel: {
-    fontSize: 12,
-    opacity: 0.7,
-    textAlign: 'center',
-  },
-  analysisDescription: {
-    fontSize: 14,
-    lineHeight: 20,
-    textAlign: 'center',
-    paddingTop: 8,
-    borderTopWidth: 1,
-    borderTopColor: 'rgba(128, 128, 128, 0.3)',
-  },
-  debugContainer: {
-    margin: 16,
-    paddingBottom: 32,
-  },
-  debugButton: {
-    padding: 12,
-    borderWidth: 1,
-    borderRadius: 8,
-    alignItems: 'center',
-  },
-  debugButtonText: {
-    fontSize: 14,
-  },
-  staleDataContainer: {
-    margin: 16,
-    padding: 16,
-    backgroundColor: 'rgba(255, 149, 0, 0.1)',
-    borderRadius: 12,
-    borderLeftWidth: 4,
-    borderLeftColor: '#FF9500',
-  },
-  staleDataText: {
-    color: '#FF9500',
-    fontSize: 16,
-    fontWeight: '600',
-    marginBottom: 4,
-  },
-  staleDataSubtext: {
-    fontSize: 14,
-    opacity: 0.8,
-  },
-  freshnessMessage: {
-    fontSize: 12,
-    opacity: 0.7,
-    marginTop: 12,
-    textAlign: 'center',
-  },
-});
